refactor : service interface 삭제
