+++ last-order-api/src/main/java/com/flab/item/controller/ItemController.java
@@ -5,7 +5,6 @@ import com.flab.item.domain.dto.response.ItemResponse;
 import com.flab.item.service.ItemService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
